move selector file and update refs
