+++ e2e/selectors/wallet/DetectedTokensView.selectors.js
